test(images): colocate edit source coverage
Keep MIME canonicalization and transcoding-deduplication tests with the shared edit-source module, and keep its MIME helper private to that module.
