+++ packages/gateway/src/data-plane/chat/responses/interceptors/server-tools/image-generation_test.ts
@@ -22,9 +22,8 @@ import {
 import { initRepo } from '../../../../../repo/index.ts';
 import { InMemoryRepo } from '../../../../../repo/memory.ts';
 import { mockChatGatewayCtx } from '../../../../../test-helpers/gateway-ctx.ts';
-import { prepareImageEditSources } from '../../../../images/edit-source.ts';
 import type { ResponsesInvocation } from '../types.ts';
-import { initExternalResourceFetcher, initImageProcessor } from '@floway-dev/platform';
+import { initExternalResourceFetcher } from '@floway-dev/platform';
 import type { CanonicalResponsesPayload, ResponsesInputImage, ResponsesInputItem, ResponsesPayload, ResponsesTool } from '@floway-dev/protocols/responses';
 import { assert, assertEquals, assertFalse, assertStringIncludes, assertThrows, stubModelCandidate } from '@floway-dev/test-utils';
 
@@ -350,30 +349,6 @@ test('inspectImageSources reads tool-result images and preserves forward order',
   assertEquals(first.mimeType, 'image/png');
   assertEquals(second.mimeType, 'image/jpeg');
   assertEquals(third.mimeType, 'image/webp');
-});
-
-test('prepareImageEditSources transcodes formats accepted by native Responses but rejected by images edits', async () => {
-  let observedTarget: unknown = undefined;
-  let processorCalls = 0;
-  initImageProcessor({
-    compressToWebp: (_input, target) => {
-      processorCalls += 1;
-      observedTarget = target;
-      return Promise.resolve(Uint8Array.of(1, 2, 3));
-    },
-  });
-  const first = {
-    bytes: Uint8Array.of(4, 5, 6).buffer,
-    mimeType: 'image/gif',
-  };
-  const sameBytes = { bytes: Uint8Array.of(4, 5, 6).buffer, mimeType: 'image/gif' };
-  const prepared = await prepareImageEditSources([first, first, sameBytes]);
-  assertEquals(processorCalls, 1);
-  assert(prepared[0] === prepared[1]);
-  assert(prepared[0] === prepared[2]);
-  assertEquals(prepared[0].mimeType, 'image/webp');
-  assertEquals([...new Uint8Array(prepared[0].bytes)], [1, 2, 3]);
-  assertEquals(observedTarget, null);
 });
 
 test('resolveImageOperation exposes a malformed replayed result as an invariant failure', () => {
